feat: update item management components for price formatting and data handling
- Enhanced CreateItem and EditItem components to format price input with Indonesian Rupiah and added live formatting features.
- Updated ListItems component to display prices formatted as Rupiah.
- Modified ItemFactory to generate prices as integers for consistency.
- Changed database schema to store prices as unsigned integers for better data integrity.
- Added new POS route to sidebar navigation for improved access to point-of-sale functionality.

+++ database/migrations/2025_12_12_174817_create_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('sku')->unique();
-            $table->decimal('price', 8, 2);
+            $table->unsignedInteger('price');
             $table->enum('status',['active','inactive'])->default('active');
             // add fields images, variations
             $table->timestamps();
